pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zenoh-10base-t1s"
version = "0.1.0"
description = "Zenoh 10BASE-T1S Automotive Master Controller Simulator"
requires-python = ">=3.9"
dependencies = [
    "eclipse-zenoh>=1.0.0",
    "claude-agent-sdk>=0.1.0",
    "pyyaml>=6.0",
    "cbor2>=5.6",
    "typer>=0.12",
    "rich>=13.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
zenoh-t1s-master = "src.master.cli:app"
sim-harness = "sim_harness.main:main"
gui-sim = "gui.run:main"

[tool.hatch.build.targets.wheel]
packages = ["src", "sim_harness", "gui"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
